emdad/requirements.txt
Django==4.2.*
python-decouple==3.8
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
phonenumbers==8.13.38
requests==2.31.0
psycopg2-binary==2.9.5
django-admin-interface==0.26.1
pytz==2024.1
